+++ src/module/utils/user-utils.ts
@@ -4,7 +4,7 @@
  * @param millis - duration to pause for in milliseconds
  * @returns A `Promise` that is resolved when the given time passed
  */
-export async function pause(millis: number): Promise<void> {
+export function pause(millis: number): Promise<void> {
 	return new Promise((resolve) => setTimeout(resolve, millis));
 }
 
@@ -14,9 +14,12 @@ export async function pause(millis: number): Promise<void> {
  * **WARNING: This will permanently delete every entity in your world (scenes, actors, items, macros, roll tables, journal entries, playlists, chat messages, folders, etc.)**
  */
 export async function clearWorld(): Promise<void> {
-	const exclude = new Set([User].map((element) => element.metadata.name));
+	const exclude = new Set([foundry.documents.User].map((element) => element.metadata.name));
 	for (const collection of Object.values(game)) {
-		if (!(collection instanceof DocumentCollection) || exclude.has(collection.documentName))
+		if (
+			!(collection instanceof foundry.documents.abstract.DocumentCollection) ||
+			exclude.has(collection.documentName)
+		)
 			continue;
 		if (collection.size === 0) continue;